h2snotifier/requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
Django==5.0.2
idna==3.6
image==1.5.33
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
python-dotenv==1.0.1
python-telegram-bot==13.7
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.1
